kidora_api/requirements.txt
# Dépendances
fastapi
uvicorn[standard]
pymongo
pandas
pydantic
python-dotenv
dnspython